added some authorization logic to configurations

+++ Modules/Blueprint/Http/Controllers/Blueprint/ConfigurationController.php
@@ -2,7 +2,7 @@
 
 namespace Modules\Blueprint\Http\Controllers\Blueprint;
 
-//use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Auth\Access\AuthorizationException;
 use App\Models\BlueprintWizardAnswer;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
@@ -18,9 +18,12 @@ class ConfigurationController extends Controller
      *
      * @param Blueprint $blueprint
      * @return View
+     * @throws AuthorizationException
      */
     public function show( Blueprint $blueprint ): View
     {
+        $this->authorize('edit_configuration', $blueprint);
+
         $configs = Configuration::where('blueprint_id', $blueprint->id )
             ->where('obsolete', false)
             ->select([
@@ -40,16 +43,22 @@ class ConfigurationController extends Controller
     /**
      * @param Blueprint $blueprint
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function reset( Blueprint $blueprint ): RedirectResponse
     {
+        $this->authorize('reset_configuration', $blueprint );
+
         Configuration::where('blueprint_id', $blueprint->id )
             ->update([
                 'value' => 0
             ]);
 
+        // clear out the selected answers so the forms are reset too.
         BlueprintWizardAnswer::where('blueprint_id', $blueprint->id)->delete();
 
-        return redirect()->route('blueprint.home', [ $blueprint ]);
+        return redirect()
+            ->route('blueprint.home', [ $blueprint ])
+            ->with('message','Successfully reset this blueprint\'s configuration');
     }
 }
